LPS-117232 Rename
