+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/internal/nativeobject/DataEngineNativeObjectTrackerImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.engine.internal.nativeobject.tracker;
+package com.liferay.data.engine.internal.nativeobject;
 
 import com.liferay.data.engine.nativeobject.DataEngineNativeObject;
 import com.liferay.data.engine.nativeobject.tracker.DataEngineNativeObjectTracker;
