+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/internal/nativeobject/DataEnginePortalExecutor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.engine.internal.nativeobject.tracker;
+package com.liferay.data.engine.internal.nativeobject;
 
 import com.liferay.petra.concurrent.NoticeableExecutorService;
 import com.liferay.petra.concurrent.ThreadPoolHandlerAdapter;
